+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [0.2.6] - 2026-01-23
-- fix #37, implement POR error
+- fix #37, implement POR error (DCTL #289)
 - minor edits
 
 ## [0.2.5] - 2026-01-02